explore the req and user decorator

+++ ecommerceBackend/src/verify-user/verify-user.service.ts
@@ -49,16 +49,18 @@ export class VerifyUserService {
         await this.userModel.deleteOne({ email: userToVerify.email });
         throw new NotAcceptableException(`Too many tries. User deleted`);
       }
-
+      console.log('testing1');
       const createdAt = new Date(userToVerify.createdAt).getTime();
       const currentTime = new Date().getTime();
       const fiveMinutesInMilliseconds = 5 * 60 * 1000;
+      console.log('testing11');
 
       if (currentTime - createdAt > fiveMinutesInMilliseconds) {
         await this.verifyUserModel.deleteOne({ email: userToVerify.email });
         await this.userModel.deleteOne({ email: userToVerify.email });
         throw new BadRequestException(`OTP expired for email : ${email}`);
       }
+      console.log('testing111');
 
       if (userToVerify.otp !== otp) {
         await this.verifyUserModel.updateOne(
@@ -67,15 +69,26 @@ export class VerifyUserService {
         );
         throw new BadRequestException(`Incorrect OTP`);
       }
+      console.log('testing1111');
 
+      console.log('test2');
       const verifyemail = await this.userModel.updateOne(
         { email: userToVerify.email },
         { is_emailVerified: true },
       );
 
+      if (!verifyemail) {
+        throw new BadRequestException(`Email not verified`);
+      }
+      const verifiedUser = await this.userModel.findOne({
+        email: userToVerify.email,
+      });
+      if (!verifiedUser) {
+        throw new BadRequestException(`User not found`);
+      }
       await this.verifyUserModel.deleteOne({ email: userToVerify.email });
 
-      return verifyemail;
+      return verifiedUser;
     } catch (error) {
       throw error;
     }
